Implement 'standardsMode' as well as the `:scope` selector
Pass an additional options array down through all of the selectors.
This allows us to test against global information (passed through
`$opts`) like the scoping root.  It also allows end users to add
custom selectors against global information (like `:target`) by
passing that in through the options array as well.

In addition, when `standardsMode` is set to `true` in the options
array, disable the various workarounds for bugs or missing features in
the PHP DOM implementation for more standards-compliant behavior.

Change-Id: I718640eb08aa4a3b551edbaab95f385d91eba9e7