Add `getElementsById` option to support multiple elements with the same ID
Some DOM implementations can efficiently index multiple elements with
the same ID, even though the standard web standards don't provide a
method to access this index (neither does the PHP dom extension).
Provide a way for the end user to pass in a callable to access this
functionality if available.

Change-Id: Iec41daeef73abed35c6f7dd70ff1bc7f3220c28b