Correctly handle tag and class names with special characters
Ensure ::getElementsByTagName() and ::getElementsByClassName() work
correctly on tag or class names that contain 'special characters' that
need to be escaped in a CSS query.

Implemented by the addition of an `encodeid` function based on
https://developer.mozilla.org/en-US/docs/Web/API/CSS/escape_static
which is used in the appropriate places, ensuring that tag and class
names with (for example) embedded colons in them get properly escaped
when they are converted back to CSS query strings.

Bug: T357812
Change-Id: I587cc1ada3de9fd3e0fa25e8054c18a5309f3ea8