Add new <mw:section> tag to test case document for better coverage
Briefly, the bug is that an element <mw:section> gets parsed as:
  prefix = "mw"
  localName = "section"
  namespaceURI = "http://www.w3.org/1999/xhtml"
and then any mutation of that element (say, by adding it to the tree or
adopting it into a DOMFragment) results in the prefix getting cleared
(presumably because "mw" isn't actually a valid prefix for the XHTML
namespace; this behavior isn't in the spec as far as I can tell but
is reproducible in Firefox) and the element "magically" becomes
  prefix = ''
  localName = 'section'
  namespaceURI = "http://www.w3.org/1999/xhtml"
aka a "plain old" <section> tag.

To remedy this, we need to force the element to be created with a
localName of "mw:section", which is (a) spec-compliant, and (b)
possible using the 'supressHtmlNamespace' option of Remex (which is
what we use in production anyway -- there are performance advantages
to this as well), and (c) possible with the standard DOM parser with a
very hacky workaround involving suppressing xml errors during the
initial parse and then re-creating the element with
$doc->createElement after the parse.

In any case, the result after all of this is a test document which
contains a standards-compliant <mw:section> tag, which we can then
reasonably run Zest tests against.

Change-Id: Id4df1b5d9ea5c15f16c291da8f2512f760f0d3c7